first major update and bugfree handling of Bitcoin option data on https://deribit.com

+++ Data/ChartData.cs
@@ -11,26 +11,26 @@ namespace MasterComputations.Data
     public class ChartData
     {
         [ProtoMember(1)]
-        public List<float> close { get; set; }
+        public List<double> close { get; set; }
         [ProtoMember(2)]
-        public List<float> cost { get; set; }
+        public List<double> cost { get; set; }
         [ProtoMember(3)]
-        public List<float> high { get; set; }
+        public List<double> high { get; set; }
         [ProtoMember(4)]
-        public List<float> low { get; set; }
+        public List<double> low { get; set; }
         [ProtoMember(5)]
-        public List<float> open { get; set; }
+        public List<double> open { get; set; }
         [ProtoMember(6)]
         public string status { get; set; }
         [ProtoMember(7)]
         public List<long> ticks { get; set; }
         [ProtoMember(8)]
-        public List<float> volume { get; set; }
+        public List<double> volume { get; set; }
 
         public ChartData()
         {
-            close = new List<float>(); cost = new List<float>(); high = new List<float>(); low = new List<float>();
-            open = new List<float>(); volume = new List<float>(); ticks = new List<long>(); status = "";
+            close = new List<double>(); cost = new List<double>(); high = new List<double>(); low = new List<double>();
+            open = new List<double>(); volume = new List<double>(); ticks = new List<long>(); status = "";
         } 
     }
 }
